+++ sqlalchemy_redshift/__init__.py
@@ -6,5 +6,5 @@ from sqlalchemy.dialects import registry
 
 registry.register("redshift", "sqlalchemy_redshift.dialect", "RedshiftDialect")
 registry.register(
-    "redshift+psycopg2", "sqlalchemy_redshift.dialect", "RedshiftDialect"
+    "redshift.psycopg2", "sqlalchemy_redshift.dialect", "RedshiftDialect"
 )