+++ sqlalchemy_redshift/commands.py
@@ -5,6 +5,7 @@ import re
 import warnings
 
 import sqlalchemy as sa
+from sqlalchemy import exc as sa_exc
 from sqlalchemy.ext import compiler as sa_compiler
 from sqlalchemy.sql import expression as sa_expression
 
@@ -113,7 +114,7 @@ class UnloadFromSelect(_ExecutableClause):
     ----------
     select: sqlalchemy.sql.selectable.Selectable
         The selectable Core Table Expression query to unload from.
-    data_location: str
+    unload_location: str
         The Amazon S3 location where the file will be created, or a manifest
         file if the `manifest` option is used
     access_key_id: str, optional
@@ -153,6 +154,12 @@ class UnloadFromSelect(_ExecutableClause):
         Overwrite the key at unload_location in the S3 bucket.
     parallel: bool, optional
         If disabled unload sequentially as one file.
+    header: bool, optional
+        Boolean value denoting whether to add header line
+        containing column names at the top of each output file.
+        Text transformation options, such as delimiter, add_quotes,
+        and escape, also apply to the header line.
+        `header` can't be used with fixed_width.
     """
 
     def __init__(self, select, unload_location, access_key_id=None,
@@ -160,11 +167,17 @@ class UnloadFromSelect(_ExecutableClause):
                  aws_account_id=None, iam_role_name=None,
                  manifest=False, delimiter=None, fixed_width=None,
                  encrypted=False, gzip=False, add_quotes=False, null=None,
-                 escape=False, allow_overwrite=False, parallel=True):
+                 escape=False, allow_overwrite=False, parallel=True,
+                 header=False):
 
         if delimiter is not None and len(delimiter) != 1:
             raise ValueError(
                 '"delimiter" parameter must be a single character'
+            )
+
+        if header and fixed_width is not None:
+            raise ValueError(
+                "'header' cannot be used with 'fixed_width'"
             )
 
         credentials = _process_aws_credentials(
@@ -179,6 +192,7 @@ class UnloadFromSelect(_ExecutableClause):
         self.unload_location = unload_location
         self.credentials = credentials
         self.manifest = manifest
+        self.header = header
         self.delimiter = delimiter
         self.fixed_width = fixed_width
         self.encrypted = encrypted
@@ -198,6 +212,7 @@ def visit_unload_from_select(element, compiler, **kw):
        UNLOAD (:select) TO :unload_location
        CREDENTIALS :credentials
        {manifest}
+       {header}
        {delimiter}
        {encrypted}
        {fixed_width}
@@ -212,6 +227,7 @@ def visit_unload_from_select(element, compiler, **kw):
 
     qs = template.format(
         manifest='MANIFEST' if el.manifest else '',
+        header='HEADER' if el.header else '',
         delimiter=(
             'DELIMITER AS :delimiter' if el.delimiter is not None else ''
         ),
@@ -267,6 +283,9 @@ class Format(enum.Enum):
     csv = 'CSV'
     json = 'JSON'
     avro = 'AVRO'
+    orc = 'ORC'
+    parquet = 'PARQUET'
+    fixed_width = 'FIXEDWIDTH'
 
 
 class Compression(enum.Enum):
@@ -311,7 +330,7 @@ class CopyCommand(_ExecutableClause):
     quote : str, optional
         Specifies the character to be used as the quote character when using
         ``format=Format.csv``. The default is a double quotation mark ( ``"`` )
-    delimiter : File delimiter, optional
+    delimiter : Field delimiter, optional
         defaults to ``|``
     path_file : str, optional
         Specifies an Amazon S3 location to a JSONPaths file to explicitly map
@@ -551,6 +570,13 @@ def visit_copy_command(element, compiler, **kw):
             value=element.path_file,
             type_=sa.String,
         ))
+    elif element.format == Format.orc:
+        format_ = 'FORMAT AS ORC'
+    elif element.format == Format.parquet:
+        format_ = 'FORMAT AS PARQUET'
+    elif element.format == Format.fixed_width and element.fixed_width is None:
+        raise sa_exc.CompileError(
+            "'fixed_width' argument required for format 'FIXEDWIDTH'.")
     else:
         format_ = ''